+++ workspace/solutions/python/src/utils/test.py
@@ -18,9 +18,8 @@ from importlib.machinery import ModuleSpec
 import json
 import pathlib
 import re
-from typing import List
 
-from .example import Example
+from .example import ExamplesSet
 
 ID_LENGTH = 4
 SOLUTION_FILE_NAME_REGEXP = re.compile(r"^s_\d{4}_\d{2}\.py$")
@@ -41,34 +40,38 @@ def t(spec: ModuleSpec) -> None:
         )
     ]
     id = "".join(paths)
-    example_path = EXAMPLES_DIRECTORY_PATH / "/".join(paths) / f"{id}.json"
-    assert example_path.is_file(), f"No example file: {example_path}"
+    examples_path = EXAMPLES_DIRECTORY_PATH / "/".join(paths) / f"{id}.json"
+    assert examples_path.is_file(), f"No example file: {examples_path}"
 
-    examples: List[Example] = json.loads(example_path.read_text())
-    assert len(examples) > 0, "No examples"
+    examples_list = json.loads(examples_path.read_text())
+    assert len(examples_list) > 0, "No examples"
+
+    examples_set_list = ExamplesSet.listOf(examples_list)
 
     # 遍历题解模块
     for solution_index, solution_file_path in enumerate(solution_file_paths):
         module = importlib.import_module(f"src.{'.'.join(solution_file_path.parts[-ID_LENGTH - 1:-1])}.{solution_file_path.stem}")
-        solution = module.Solution()
-        methods = [method_name for method_name in dir(solution) if callable(getattr(solution, method_name)) and not method_name.startswith("_")]
-        assert len(methods) > 0, "No solution method"
-        assert len(methods) == 1, "Too many solution methods"
-        method = getattr(solution, methods[0])
+        for examples_index, examples_set in enumerate(examples_set_list):
+            solution = module.Solution(*examples_set.init)
+            methods = [method_name for method_name in dir(solution) if callable(getattr(solution, method_name)) and not method_name.startswith("_")]
+            assert len(methods) > 0, "No solution method"
+            assert len(methods) == 1, "Too many solution methods"
+            method = getattr(solution, methods[0])
 
-        # 遍历示例
-        for example_index, example in enumerate(examples):
-            input = example["input"]
-            output = example["output"]
-            result = method(*input)
-            assert result == output, "\n".join(
-                [
-                    "",
-                    f"solution: {solution_index}",
-                    f"example:  {example_index}",
-                    f"input:    {input}",
-                    f"result:   {result}",
-                    f"expected: {output}",
-                    "",
-                ]
-            )
+            # 遍历示例
+            for example_index, example in enumerate(examples_set.examples):
+                input = example.input
+                output = example.output
+                result = method(*input)
+                assert result == output, "\n".join(
+                    [
+                        "",
+                        f"solution: {solution_index}",
+                        f"examples: {examples_index}",
+                        f"example:  {example_index}",
+                        f"input:    {input}",
+                        f"result:   {result}",
+                        f"expected: {output}",
+                        "",
+                    ]
+                )